requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.1
certifi==2021.10.8
click==8.1.3
colorama==0.4.4
elastic-transport==8.1.2
elasticsearch==8.2.0
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
